theme updates , db integrations
slowly building the app

+++ app/src/main/java/com/refupanker/teachmetech/model/mdl_user.kt
@@ -5,7 +5,6 @@ import java.io.Serializable
 data class mdl_user(
     val token: String,
     val name: String,
-    val rank: Int = 0,
+    val rank: Long = 0,
     val active: Boolean = true,
-
     ) : Serializable
